refactor : UserRepository 파일 디렉터리 위치 수정
